refacor: rename
- rename dispatch/create -> index
- rename store/create -> index
