+++ src/view/hiddenselect.tsx
@@ -1,6 +1,6 @@
 import { h, Component } from 'preact';
 import { connect } from './hoc';
-import { DispatchProps } from '../dispatch/create';
+import { DispatchProps } from '../dispatch';
 import { Data } from '@t/store/data';
 import { cls } from '../css/constants';
